--- src/main/java/com/onenth/OneNth/domain/alert/keywordAlert/entity/enums/KeywordAlertType.java
@@ -1,5 +1,0 @@
-package com.onenth.OneNth.domain.alert.keywordAlert.entity.enums;
-
-public enum KeywordAlertType {
-    REGION, PRODUCT
-}
