--- src/main/java/com/onenth/OneNth/domain/alert/keywordAlert/service/KeywordAlertQueryService.java
@@ -1,4 +1,0 @@
-package com.onenth.OneNth.domain.alert.keywordAlert.service;
-
-public interface KeywordAlertQueryService {
-}
