--- src/main/java/com/onenth/OneNth/domain/alert/keywordAlert/service/KeywordAlertQueryServiceImpl.java
@@ -1,4 +1,0 @@
-package com.onenth.OneNth.domain.alert.keywordAlert.service;
-
-public class KeywordAlertQueryServiceImpl {
-}
