+++ src/main/java/com/onenth/OneNth/domain/member/entity/enums/KeywordAlertType.java
@@ -1,0 +1,5 @@
+package com.onenth.OneNth.domain.member.entity.enums;
+
+public enum KeywordAlertType {
+    REGION, PRODUCT
+}
