+++ src/main/java/com/onenth/OneNth/domain/member/settings/alert/keywordAlert/converter/KeywordAlertConverter.java
@@ -1,9 +1,9 @@
-package com.onenth.OneNth.domain.alert.keywordAlert.converter;
+package com.onenth.OneNth.domain.member.settings.alert.keywordAlert.converter;
 
-import com.onenth.OneNth.domain.alert.keywordAlert.dto.KeywordAlertResponseDTO;
-import com.onenth.OneNth.domain.alert.keywordAlert.entity.ProductKeywordAlert;
-import com.onenth.OneNth.domain.alert.keywordAlert.entity.RegionKeywordAlert;
-import com.onenth.OneNth.domain.alert.keywordAlert.entity.enums.KeywordAlertType;
+import com.onenth.OneNth.domain.member.settings.alert.keywordAlert.dto.KeywordAlertResponseDTO;
+import com.onenth.OneNth.domain.member.entity.ProductKeywordAlert;
+import com.onenth.OneNth.domain.member.entity.RegionKeywordAlert;
+import com.onenth.OneNth.domain.member.entity.enums.KeywordAlertType;
 import com.onenth.OneNth.domain.member.entity.Member;
 import com.onenth.OneNth.domain.region.entity.Region;
