+++ src/main/java/com/onenth/OneNth/domain/member/settings/alert/keywordAlert/dto/KeywordAlertRequestDTO.java
@@ -1,4 +1,4 @@
-package com.onenth.OneNth.domain.alert.keywordAlert.dto;
+package com.onenth.OneNth.domain.member.settings.alert.keywordAlert.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
