+++ src/main/java/com/onenth/OneNth/domain/member/settings/alert/keywordAlert/repository/RegionKeywordAlertRepository.java
@@ -1,6 +1,6 @@
-package com.onenth.OneNth.domain.alert.keywordAlert.repository;
+package com.onenth.OneNth.domain.member.settings.alert.keywordAlert.repository;
 
-import com.onenth.OneNth.domain.alert.keywordAlert.entity.RegionKeywordAlert;
+import com.onenth.OneNth.domain.member.entity.RegionKeywordAlert;
 import com.onenth.OneNth.domain.member.entity.Member;
 import com.onenth.OneNth.domain.region.entity.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
