+++ src/main/java/com/onenth/OneNth/domain/member/settings/alert/keywordAlert/service/KeywordAlertQueryService.java
@@ -1,0 +1,4 @@
+package com.onenth.OneNth.domain.member.settings.alert.keywordAlert.service;
+
+public interface KeywordAlertQueryService {
+}
