+++ src/main/java/com/onenth/OneNth/domain/member/settings/alert/keywordAlert/service/KeywordAlertQueryServiceImpl.java
@@ -1,0 +1,4 @@
+package com.onenth.OneNth.domain.member.settings.alert.keywordAlert.service;
+
+public class KeywordAlertQueryServiceImpl {
+}
